update script arguments implementation
