+++ demisto_sdk/commands/content_graph/objects/base_script.py
@@ -18,6 +18,7 @@ from demisto_sdk.commands.content_graph.common import (
     RelationshipType,
     lazy_property,
 )
+from functools import cached_property
 from demisto_sdk.commands.content_graph.objects.base_content import (
     BaseNode,
 )
@@ -27,20 +28,15 @@ from demisto_sdk.commands.content_graph.objects.integration_script import (
 from demisto_sdk.commands.prepare_content.preparers.marketplace_incident_to_alert_scripts_prepare import (
     MarketplaceIncidentToAlertScriptsPreparer,
 )
-
-
-@dataclass
-class Argument:
-    name: str
-    description: str
-    deprecated: bool = False
-    default_value: Optional[str] = None
+from pydantic import Field
+from demisto_sdk.commands.content_graph.parsers.base_script import Argument
 
 
 class BaseScript(IntegrationScript, content_type=ContentType.BASE_SCRIPT):  # type: ignore[call-arg]
     tags: List[str]
     skip_prepare: List[str]
     runas: str = ""
+    arguments: List[Argument] = Field(exclude=False)
 
     def metadata_fields(self) -> Set[str]:
         return (
@@ -69,18 +65,6 @@ class BaseScript(IntegrationScript, content_type=ContentType.BASE_SCRIPT):  # ty
             data["nativeimage"] = supported_native_images
 
         return data
-
-    @lazy_property
-    def arguments(self) -> List[Argument]:
-        return [
-            Argument(
-                name=argument.get("name", ""),
-                deprecated=argument.get("deprecated", False),
-                description=argument.get("description", ""),
-                default_value=argument.get("defaultValue"),
-            )
-            for argument in get_yaml(self.path).get("args", [])
-        ]
 
     @property
     def imported_by(self) -> List[BaseNode]:
